+++ spec/generators/kafka_consumer_generator_spec.rb
@@ -1,19 +1,17 @@
 # frozen_string_literal: true
 
-require 'generators/deimos/active_record_generator'
+require 'generators/deimos/kafka_consumer_generator'
 
 RSpec.describe Deimos::Generators::KafkaConsumerGenerator do
 
   before(:each) do
-    # FileUtils.touch('configure.rb')
-    # FileUtils.rm_rf('db') if File.exist?('db')
-    # FileUtils.rm_rf('app') if File.exist?('app')
+    FileUtils.touch('configure.rb')
   end
 
   after(:each) do
     FileUtils.rm_rf('db') if File.exist?('db')
     FileUtils.rm_rf('app') if File.exist?('app')
-    # FileUtils.rm('configure.rb') if File.exist?('configure.rb')
+    FileUtils.rm('configure.rb') if File.exist?('configure.rb')
   end
 
   it 'should generate a migration' do
@@ -58,7 +56,28 @@ RSpec.describe Deimos::Generators::KafkaConsumerGenerator do
         enum an_enum: {sym1: 'sym1', sym2: 'sym2'}
       end
     MODEL
+    consumer = <<~consumer
+      class GeneratedTableConsumer < Deimos::Consumer
+      
+        # Consume the incoming topic
+        # @param payload [Hash]
+        # @param metadata [Hash]
+        def consume(payload, metadata)
+        end
+      end
+    consumer
+    consumer_config_block = <<~consumer_config_block
+      consumer do
+        topic 'GeneratedTopic'
+        class_name 'GeneratedTableConsumer'
+        namespace 'com.my-namespace'
+        schema 'Generated'
+        group_id 'PLEASE FILL THIS'
+      end
+    consumer_config_block
     expect(File.read('app/models/generated_table.rb')).to eq(model)
+    expect(File.read('app/lib/kafka/generated_table_consumer.rb')).to eq(consumer)
+    expect(File.read('configure.rb')).to eq(consumer_config_block)
   end
 
 end